updated reservation

+++ src/main/java/org/example/model/ReservationService.java
@@ -1,5 +1,6 @@
 package org.example.model;
 
+import org.example.controller.ReservationValidator;
 import org.example.model.builder.GroupReservationBuilder;
 import org.example.model.builder.ReservationBuilder;
 import org.example.model.builder.StandardReservationBuilder;
@@ -9,9 +10,10 @@ import org.example.model.dao.TableDAO;
 import org.example.model.entities.Customer;
 import org.example.model.entities.Reservation;
 import org.example.model.entities.Table;
-import org.example.strategy.GroupAllocationStrategy;
-import org.example.strategy.StandardAllocationStrategy;
-import org.example.strategy.TableAllocationStrategy;
+import org.example.strategy.reservation_display.ReservationDisplayStrategy;
+import org.example.strategy.table_allocation.GroupAllocationStrategy;
+import org.example.strategy.table_allocation.StandardAllocationStrategy;
+import org.example.strategy.table_allocation.TableAllocationStrategy;
 
 import java.time.LocalDateTime;
 import java.util.*;
@@ -21,6 +23,9 @@ public class ReservationService {
     private CustomerDAO customerDAO;
     private TableDAO tableDAO;
     private TableAllocationStrategy allocationStrategy;
+    private ReservationValidator reservationValidator = new ReservationValidator();
+
+    private ReservationDisplayStrategy displayStrategy;
 
 
     public ReservationService(ReservationDAO reservationDAO, TableDAO tableDAO, CustomerDAO customerDAO) {
@@ -32,8 +37,15 @@ public class ReservationService {
     public void setAllocationStrategy(TableAllocationStrategy allocationStrategy) {
         this.allocationStrategy = allocationStrategy;
     }
+    public void setDisplayStrategy(ReservationDisplayStrategy displayStrategy) {
+        this.displayStrategy = displayStrategy;
+    }
 
-    public void addReservation(Map<String, Object> parameters) {
+    public String addReservation(Map<String, Object> parameters) {
+        if (!reservationValidator.validateReservationParameters(parameters)) {
+            return "Validation failed for reservation parameters.";
+        }
+
         Customer customer = customerDAO.getCustomerByName((String) parameters.get("customerName"));
         if (customer == null) {
             customer = new Customer();
@@ -50,28 +62,43 @@ public class ReservationService {
         int peopleTotal = (int) parameters.get("peopleTotal");
         List<Table> availableTables = tableDAO.getAvailableTables();
         List<Integer> suitableTableIds = allocationStrategy.allocateTables(availableTables, peopleTotal);
+        if (suitableTableIds.isEmpty()) {
+            return "Failed to find suitable tables for the reservation.";
+        }
 
         ReservationBuilder builder = isGroup ? new GroupReservationBuilder() : new StandardReservationBuilder();
         builder.setCustomerId(customer.getId())
                 .setStartTime((LocalDateTime) parameters.get("startTime"))
                 .setEndTime((LocalDateTime) parameters.get("endTime"))
-                .setTableIds(suitableTableIds);
+                .setTableIds(suitableTableIds)
+                .setCustomerName(customer.getName());
         if (isGroup) {
             builder.setSpecialSetup((String) parameters.get("specialSetup"));
         }
 
         Reservation reservation = builder.build();
         reservationDAO.insertReservation(reservation, suitableTableIds);
-    }
+        return "Reservation added successfully.";    }
 
-    public void deleteReservation(int reservationId) {
+    public String deleteReservation(int reservationId) {
+        if(reservationId < 0) {
+            return "Couldn't delete reservation. Try again.";
+        }
         reservationDAO.delete(reservationId);
+        return "Reservation deleted successfully.";
     }
 
     public List<Reservation> getAllReservations() {
         return reservationDAO.getAll();
     }
 
+    public List<Reservation> executeDisplayStrategy() {
+        if (displayStrategy == null) {
+            throw new IllegalStateException("Display strategy not set");
+        }
+        List<Reservation> allReservations = getAllReservations();
+        return displayStrategy.execute(allReservations);
+    }
 }
